Move cachito_util.py into utils/
*OSBS-8605

Also change files which import cachito_util.py
