+++ atomic_reactor/plugins/pre_reactor_config.py
@@ -9,7 +9,7 @@ of the BSD license. See the LICENSE file for details.
 from __future__ import absolute_import
 
 from copy import deepcopy
-from atomic_reactor.cachito_util import CachitoAPI
+from atomic_reactor.utils.cachito import CachitoAPI
 from atomic_reactor.plugin import PreBuildPlugin
 from atomic_reactor.constants import (CONTAINER_BUILD_METHODS, CONTAINER_DEFAULT_BUILD_METHOD,
                                       CONTAINER_BUILDAH_BUILD_METHOD)
